requirements.txt
# Core dependencies
streamlit==1.28.1
pandas==2.1.3
numpy==1.26.2

# Text extraction
pdfminer.six==20221105
python-docx==1.1.0
docx2txt==0.8

# NLP and embeddings
spacy==3.7.2
sentence-transformers==2.2.2
transformers==4.35.2

# ML and ranking
scikit-learn==1.3.2
xgboost==2.0.2

# API (optional)
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6

# LLM integrations (optional)
openai==1.3.7
anthropic==0.7.7

# Utilities
python-dotenv==1.0.0
tqdm==4.66.1
joblib==1.3.2

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Jupyter notebook
jupyter==1.0.0
notebook==7.0.6
ipykernel==6.27.1
matplotlib==3.8.2
seaborn==0.13.0